fix: verifiedAt Double→Int/Long, align error codes 9010-9014 across all SDKs
- Swift: IdentityAttestation.verifiedAt Double → Int (source + tests)
- Kotlin: IdentityAttestation.verifiedAt Double → Long, .toDouble() → .toLong() (source + tests)
- TS/Swift/Kotlin: Renumber SCP-ATTEST-9001..9005 → 9010..9014 to avoid
  collision with Android's SCP-ATTEST-9001 (Play Integrity). Aligns with
  Python SDK which was already renumbered.

+++ bindings/kotlin/scp-kt/src/main/kotlin/works/limn/scp/Identity.kt
@@ -325,7 +325,7 @@ data class IdentityAttestation(
     val platform: String,
     val platformHandle: String,
     val verificationMethod: String,
-    val verifiedAt: Double,
+    val verifiedAt: Long,
     val revocationStatus: RevocationStatus = RevocationStatus.Active,
     val platformId: String? = null,
 ) {
@@ -349,7 +349,7 @@ data class IdentityAttestation(
                 platform = obj["platform"]!!.jsonPrimitive.content,
                 platformHandle = obj["platform_handle"]!!.jsonPrimitive.content,
                 verificationMethod = obj["verification_method"]!!.jsonPrimitive.content,
-                verifiedAt = obj["verified_at"]!!.jsonPrimitive.content.toDouble(),
+                verifiedAt = obj["verified_at"]!!.jsonPrimitive.content.toLong(),
                 revocationStatus = revocationStatus,
                 platformId = obj["platform_id"]?.jsonPrimitive?.content,
             )
